feat(plugin): add `intermediate` old interface
Added `intermediate` old depecated interface
for compatibility with `gulp-intermediate`
and tests for it
